test: testing favorites service

+++ services/FavoriteService.js
@@ -8,8 +8,7 @@ const Favorite = mongoose.model('Favorite',FavoriteSchema)
 module.exports.FavoriteService = class FavoriteService{
     static addOneFavorite = async function(user_id, place_id, options, callback){
         try{
-
-            if(user_id && mongoose.isValidObjectId(user_id), place_id && mongoose.isValidObjectId(place_id)){
+            if(user_id && mongoose.isValidObjectId(user_id) && place_id && mongoose.isValidObjectId(place_id)){
                 const new_favorite = new Favorite({
                     user: user_id,
                     place: place_id,
@@ -20,24 +19,29 @@ module.exports.FavoriteService = class FavoriteService{
                     callback(err)
                 }else{
                     await new_favorite.save()
-                    callback(null, new_favorite)
+                    callback(null, new_favorite.toObject())
                 }
             }else{
-                ErrorGenerator.controlIntegrityofID({user_id, place_id})
+                const err = ErrorGenerator.controlIntegrityofID({user_id, place_id})
+                callback({msg : err, type_error:"no-valid"})
             }
         }catch(e){
-            callback({msg: "error in server", e, type_error:"error-mongo"})
+            if(e.code === 11000){
+                callback({msg:"vous avez déjà ce lieu en favoris", type_error:"no-valid"})
+            }else{
+                callback({msg: "error in server", e, type_error:"error-mongo"})
+            }
         }
     }
 
-    static updateOneFavorite = async function(place_id, user_id,update,options, callback){
+    static updateOneFavorite = async function(favorite_id,update,options, callback){
         try{
-            if(user_id && mongoose.isValidObjectId(user_id), place_id && mongoose.isValidObjectId(place_id)){
+            if(favorite_id && mongoose.isValidObjectId(favorite_id)){
                 if(update){
-                    Favorite.findOneAndUpdate({user:user_id, place: place_id},update,{returnDocument: 'after', runValidators: true}).then((value)=>{
+                    Favorite.findByIdAndUpdate(favorite_id,update,{returnDocument: 'after', runValidators: true}).then((value)=>{
                         try{
                             if(value){
-                                callback(null, value)
+                                callback(null, value.toObject())
                             }else{
                                 callback({msg: "Favoris non trouvée", fields_with_error: [], fields:"", type_error:"no-found"})
                             }
@@ -52,7 +56,8 @@ module.exports.FavoriteService = class FavoriteService{
                     callback({msg: "update is missing", type_error:"no-valid"})
                 }
             }else{
-                ErrorGenerator.controlIntegrityofID({user_id, place_id})
+                const err = ErrorGenerator.controlIntegrityofID({favorite_id})
+                callback({msg : err, type_error:"no-valid"})
             }
         }catch(e){
             callback({msg: "error in server", e, type_error:"error-mongo"})
@@ -60,10 +65,10 @@ module.exports.FavoriteService = class FavoriteService{
     }
 
     static deleteOneFavorite = async function(place_id, user_id, options, callback){
-        if(user_id && mongoose.isValidObjectId(user_id), place_id && mongoose.isValidObjectId(place_id)){
+        if(user_id && mongoose.isValidObjectId(user_id) && place_id && mongoose.isValidObjectId(place_id)){
                 Favorite.findOneAndDelete({place:place_id, user:user_id}).then((value)=>{
                     if(value){
-                        callback(null, value)
+                        callback(null, value.toObject())
                     }else{
                         callback({msg:"Aucun favoris trouvé", type_error:"no-found"})
                     }
@@ -72,13 +77,14 @@ module.exports.FavoriteService = class FavoriteService{
                 })
 
         }else{
-            ErrorGenerator.controlIntegrityofID({user_id, place_id})
+            const err = ErrorGenerator.controlIntegrityofID({user_id, place_id})
+            callback({msg : err, type_error:"no-valid"})
         }
     }
 
-    static deleteManyFavorite = async function(place_id, user_id, options, callback){
-        if(user_id && mongoose.isValidObjectId(user_id), place_id && mongoose.isValidObjectId(place_id)){
-                Favorite.deleteMany({place:place_id, user:user_id}).then((value)=>{
+    static deleteManyFavorites = async function(placeOrUser_id, options, callback){
+        if(placeOrUser_id && mongoose.isValidObjectId(placeOrUser_id)){
+                Favorite.deleteMany({$or:[{place:placeOrUser_id}, {user:placeOrUser_id}]}).then((value)=>{
                     if(value && value.deletedCount !== 0){
                         callback(null, value)
                     }else{
@@ -89,7 +95,8 @@ module.exports.FavoriteService = class FavoriteService{
                 })
 
         }else{
-            ErrorGenerator.controlIntegrityofID({user_id, place_id})
+            const err = ErrorGenerator.controlIntegrityofID({placeOrUser_id})
+            callback({msg : err, type_error:"no-valid"})
         }
     }
 }
